+++ src/app/modules/shared/components/list-view/list-view.component.html
@@ -1,0 +1,44 @@
+<ng-container *ngIf="!(isFetchingObservations$ | async); else skeleton">
+  <ng-container *ngIf="registrations$ | async as registrations">
+    <ion-list class="loaded" *ngIf="registrations.length > 0; else empty">
+      <app-observation-list-card *ngFor="let obs of registrations; trackBy: trackByIdFunc" [obs]="obs">
+      </app-observation-list-card>
+    </ion-list>
+    <ion-infinite-scroll (ionInfinite)="loadNextPage()" [disabled]="shouldDisableScroller$ | async">
+      <ion-infinite-scroll-content> </ion-infinite-scroll-content>
+    </ion-infinite-scroll>
+    <ion-row *ngIf="(searchResult.maxItemsFetched$ | async) === true" class="full-grid-row max-count-reached">
+      <ion-col class="ion-text-center ion-margin-horizontal">
+        <h2>{{ "OBSERVATION_LIST.MAX_COUNT_REACHED_HEADER" | translate }}</h2>
+        <h3 class="ion-text-wrap">
+          {{ "OBSERVATION_LIST.MAX_COUNT_REACHED_TEXT" | translate: { maxCount: maxCount } }}
+        </h3>
+      </ion-col>
+    </ion-row>
+  </ng-container>
+</ng-container>
+
+<ng-template #skeleton>
+  <app-observation-skeleton></app-observation-skeleton>
+  <app-observation-skeleton></app-observation-skeleton>
+  <app-observation-skeleton></app-observation-skeleton>
+</ng-template>
+
+<ng-template #empty>
+  <ion-grid class="full-grid grid-absolute">
+    <ion-row class="full-grid-row">
+      <div class="center">
+        <svg-icon class="rectangle" src="/assets/images/empty-states/rectangle.svg"></svg-icon>
+        <svg-icon class="pin1" src="/assets/images/empty-states/pin.svg"></svg-icon>
+        <svg-icon class="pin2" src="/assets/images/empty-states/pin.svg"></svg-icon>
+        <svg-icon class="pin3" src="/assets/images/empty-states/pin.svg"></svg-icon>
+      </div>
+    </ion-row>
+    <ion-row class="col-flex-direction ion-text-center ion-margin-horizontal">
+      <h2>{{ "OBSERVATION_LIST.NO_OBSERVATIONS" | translate }}</h2>
+    </ion-row>
+    <ion-row class="col-flex-direction ion-text-center ion-margin-horizontal">
+      <h3 class="ion-text-wrap">{{ "OBSERVATION_LIST.NO_OBSERVATIONS_TEXT" | translate }}</h3></ion-row
+    >
+  </ion-grid>
+</ng-template>
