+++ src/app/modules/shared/components/list-view/list-view.component.scss
@@ -1,0 +1,69 @@
+.grid-absolute {
+  position: absolute;
+  top: 0;
+  background-color: #dcecf2;
+  z-index: 2;
+
+  ion-row {
+    justify-content: center;
+    align-items: center;
+  }
+}
+
+virtual-scroller,
+ion-list {
+  opacity: 0;
+  transition: opacity 1s linear;
+
+  &.loaded {
+    opacity: 1;
+  }
+}
+
+.center {
+  width: 300px;
+  height: 300px;
+  display: inline-block;
+  position: relative;
+
+  .rectangle {
+    position: absolute;
+    top: 50%;
+    left: 50%;
+    transform: translate(-50%, -50%);
+  }
+
+  .pin1 {
+    position: absolute;
+    right: 30%;
+    bottom: 23%;
+  }
+
+  .pin2 {
+    position: absolute;
+    left: 25%;
+    top: 10%;
+  }
+
+  .pin3 {
+    position: absolute;
+    left: 11%;
+    top: 35%;
+  }
+}
+
+@media only screen and (min-width: 700px) {
+  .center {
+    margin-top: 20vh;
+  }
+}
+
+@media only screen and (max-width: 699px) {
+  .center {
+    margin-top: 5vh;
+  }
+}
+
+.max-count-reached {
+  padding-bottom: 50px;
+}
