+++ src/app/pages/modal-pages/fullscreen-image-modal/fullscreen-image-modal.page.html
@@ -26,7 +26,9 @@
     <div class="img-frame">
       <img [src]="allImages[activeImageIndex].Url || allImages[activeImageIndex]" (click)="closeModal()" />
       <ion-item *ngIf="href as href" class="link-title">
-        <a [href]="href.url" target="_blank"> {{ href.title | translate }}</a>
+        <a [href]="href.url" target="_blank">
+          <span>{{ href.title | translate }}</span> <img src="../../../../assets/icon/newTab.svg" alt="new tab"
+        /></a>
       </ion-item>
     </div>
     <button *ngIf="!isFirstSlide" class="swiper-button-prev swiper-button-custom" (click)="prevSlide()"></button>
