+++ src/app/pages/observation-list/observation-list.page.html
@@ -26,21 +26,23 @@
         <!--titelen -->
       </ion-col>
       <ion-col>
-        <span *ngIf="!isNative" class="button-container">
-          <span>{{'OBSERVATION_LIST.SORT_BY_VIEW' | translate}}:</span>
-          <ion-button
+        <div *ngIf="!isNative" class="button-container">
+          <span>{{'OBSERVATION_LIST.VIEW_TYPE' | translate}}:</span>
+          <button
             id="list"
             [ngClass]="viewType$.getValue() == 'list' && 'active'"
             (click)="changeViewType($event.currentTarget.id)"
-            ><img alt="list view" src="../../../assets/icon/reorder.svg"
-          /></ion-button>
-          <ion-button
+          >
+            <img alt="list view" src="../../../assets/icon/reorder.svg" />
+          </button>
+          <button
             id="grid"
             [ngClass]="viewType$.getValue() == 'grid' && 'active'"
             (click)="changeViewType($event.currentTarget.id)"
-            ><img alt="images view" src="../../../assets/icon/grid_view.svg"
-          /></ion-button>
-        </span>
+          >
+            <img alt="images view" src="../../../assets/icon/grid_view.svg" />
+          </button>
+        </div>
       </ion-col>
       <ion-col class="order-by-select">
         <ng-container *ngIf="orderBy$ | async as orderBy">
@@ -60,55 +62,7 @@
     </ion-row>
   </ion-grid>
   <app-images-grid *ngIf="viewType$.getValue()==='grid'" [attachmentsResult]="attachmentsResult"></app-images-grid>
-  <!--LIST VIEW START-->
-  <!--We could place it in a separate component-->
-  <ng-container *ngIf="viewType$.getValue()==='list'">
-    <ng-container *ngIf="!(isFetchingObservations$ | async); else skeleton">
-      <ng-container *ngIf="registrations$ | async as registrations">
-        <ion-list class="loaded" *ngIf="(registrations.length > 0) else empty">
-          <app-observation-list-card *ngFor="let obs of registrations; trackBy: trackByIdFunc;" [obs]="obs">
-          </app-observation-list-card>
-        </ion-list>
-        <ion-infinite-scroll (ionInfinite)="loadNextPage()" [disabled]="shouldDisableScroller$ | async">
-          <ion-infinite-scroll-content> </ion-infinite-scroll-content>
-        </ion-infinite-scroll>
-        <ion-row *ngIf="(searchResult.maxItemsFetched$ | async) === true" class="full-grid-row max-count-reached">
-          <ion-col class="ion-text-center ion-margin-horizontal">
-            <h2>{{ 'OBSERVATION_LIST.MAX_COUNT_REACHED_HEADER' | translate }}</h2>
-            <h3 class="ion-text-wrap">
-              {{ 'OBSERVATION_LIST.MAX_COUNT_REACHED_TEXT' | translate: {maxCount: maxCount} }}
-            </h3>
-          </ion-col>
-        </ion-row>
-      </ng-container>
-    </ng-container>
-  </ng-container>
-  <!--LIST VIEW END-->
-
-  <ng-template #skeleton>
-    <app-observation-skeleton></app-observation-skeleton>
-    <app-observation-skeleton></app-observation-skeleton>
-    <app-observation-skeleton></app-observation-skeleton>
-  </ng-template>
+  <app-list-view *ngIf="viewType$.getValue()==='list'" [searchResult]="searchResult"></app-list-view>
   <app-geo-fab slot="fixed"></app-geo-fab>
 </ion-content>
 <app-add-menu></app-add-menu>
-
-<ng-template #empty>
-  <ion-grid class="full-grid grid-absolute">
-    <ion-row class="full-grid-row">
-      <div class="center">
-        <svg-icon class="rectangle" src="/assets/images/empty-states/rectangle.svg"></svg-icon>
-        <svg-icon class="pin1" src="/assets/images/empty-states/pin.svg"></svg-icon>
-        <svg-icon class="pin2" src="/assets/images/empty-states/pin.svg"></svg-icon>
-        <svg-icon class="pin3" src="/assets/images/empty-states/pin.svg"></svg-icon>
-      </div>
-    </ion-row>
-    <ion-row class="col-flex-direction ion-text-center ion-margin-horizontal">
-      <h2>{{ 'OBSERVATION_LIST.NO_OBSERVATIONS' | translate }}</h2>
-    </ion-row>
-    <ion-row class="col-flex-direction ion-text-center ion-margin-horizontal">
-      <h3 class="ion-text-wrap">{{'OBSERVATION_LIST.NO_OBSERVATIONS_TEXT'|translate}}</h3></ion-row
-    >
-  </ion-grid>
-</ng-template>
