+++ src/app/pages/observation-list/observation-list.page.scss
@@ -6,18 +6,6 @@
   bottom: 0;
 }
 
-.grid-absolute {
-  position: absolute;
-  top: 0;
-  background-color: #dcecf2;
-  z-index: 2;
-
-  ion-row {
-    justify-content: center;
-    align-items: center;
-  }
-}
-
 .select-wrapper {
   display: flex;
   align-items: center;
@@ -52,13 +40,8 @@ ion-grid {
 .filter-panel {
   margin: 0 1rem;
 
-  ion-button {
-    --background: none;
-    --box-shadow: none;
-  }
-
   .active {
-    --background: #fff;
+    background: #fff;
     pointer-events: none;
   }
 }
@@ -108,76 +91,18 @@ ion-item-divider {
   }
 }
 
-@media only screen and (min-width: 700px) {
-  .center {
-    margin-top: 20vh;
-  }
-}
-
-@media only screen and (max-width: 699px) {
-  .center {
-    margin-top: 5vh;
-  }
-}
-
 .button-container {
   margin-right: 1rem;
   display: flex;
   align-items: center;
   gap: 5px;
+  button {
+    width: 40px;
+    height: 35px;
+    background-color: transparent;
+    border-radius: 5px;
+  }
   img {
-    width: 100%;
+    width: 60%;
   }
-}
-
-virtual-scroller,
-ion-list {
-  // width: 100%;
-  // height: 100%;
-  opacity: 0;
-  transition: opacity 1s linear;
-
-  &.loaded {
-    opacity: 1;
-  }
-}
-
-.last-item {
-  height: 50px;
-}
-
-.center {
-  width: 300px;
-  height: 300px;
-  display: inline-block;
-  position: relative;
-
-  .rectangle {
-    position: absolute;
-    top: 50%;
-    left: 50%;
-    transform: translate(-50%, -50%);
-  }
-
-  .pin1 {
-    position: absolute;
-    right: 30%;
-    bottom: 23%;
-  }
-
-  .pin2 {
-    position: absolute;
-    left: 25%;
-    top: 10%;
-  }
-
-  .pin3 {
-    position: absolute;
-    left: 11%;
-    top: 35%;
-  }
-}
-
-.max-count-reached {
-  padding-bottom: 50px;
 }
