+++ src/assets/i18n/nb.json
@@ -254,17 +254,17 @@
     "FORMS": "Skjemaer",
     "MAX_COUNT_REACHED_HEADER": "Ingen flere observasjoner",
     "MAX_COUNT_REACHED_TEXT": "Vi viser maks {{ maxCount }} observasjoner her. Hvis du ikke fant det du lette etter, prøv å zoome inn på kartet eller snevre inn perioden du vil se observasjoner fra.",
-    "NO_IMAGES": "Ingen vedlegg",
+    "NO_IMAGES": "Ingen bilder",
     "NO_IMAGES_TEXT": "Det er ingen observasjoner med bilder i kartutsnittet. Forsøk å endre utsnittet eller utvid perioden du vil se observasjoner fra.",
     "NO_OBSERVATIONS": "Ingen observasjoner",
     "NO_OBSERVATIONS_TEXT": "Det er ingen observasjoner i kartutsnittet, forsøk å endre utsnittet eller utvid perioden du vil se observasjoner fra.",
     "SORT_BY_CHANGE_TIME": "Oppdatert tid",
     "SORT_BY_OBS_TIME": "Observert tid",
     "SORT_BY_TEXT": "Sorter på",
-    "SORT_BY_VIEW": "Visning",
     "TITLE": "Observasjoner",
     "TOGGLE_BUTTON_OBSERVATIONS_ALL": "Alle",
-    "TOGGLE_BUTTON_OBSERVATIONS_IN_MAP_VIEW": "I kartutsnitt"
+    "TOGGLE_BUTTON_OBSERVATIONS_IN_MAP_VIEW": "I kartutsnitt",
+    "VIEW_TYPE": "Visning"
   },
   "OFFLINE_MAP": {
     "DELETE_PACKAGE_CONFIRM": "Vil du slette denne kartpakka?",
@@ -425,7 +425,7 @@
       "TAKE_NEW_PHOTO": "Ta et nytt bilde",
       "TITLE": "Bilde og notat"
     },
-    "GO_TO": "Gå til observasjonen",
+    "GO_TO": "Gå til en observasjon",
     "GROUP": {
       "SUBTITLE": "Del observasjonen med",
       "TITLE": "Del med gruppe"
